requirements.txt
absl-py==0.13.0
astunparse==1.6.3
autopep8==1.5.7
bidict==0.21.2
biosppy==0.7.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.0.0
google-cloud-translate==3.3.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
gTTS==2.2.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.16
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
Pillow==8.3.1
playsound==1.3.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-vlc==3.0.12118
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.62.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
